Set resource property fix and additional tests. (#5)
* Change export action to create application properties file with UCD version. Refactored UCADF loop actions and added paging loop action.

* Added UcdTeamRole class. Added method to convert Properties collection to UcdProperty list.

* Chaange type of run application start date. Accrue outprops for actions performed by runner.

* Added actionInfo to UcdSendMail. Added process request type enum.

* Ability to get resource properties by list or map. Actions runner now interprets map key variables. Ability to get snapshot versions as a list or map. Added UcAdfExit and UcdUpdateComponent actions.

* Fixes to UcdFindResources.

* Fixed ucadfclient variables. Added jaxb libraries needed for newer Java versions.

* Added rename resource action and delete agent retry.

* Fix set resource security property. Additional tests.

* Updated tests.

* Added more actions, tests, and package management capabilities.

* Added delete authentication realm, delete authorization realm actions, get statuses actions.
Added post-processing script actions.
Added get system properties and delete system property action.
Fixed HTTP select import problems.
Updated jackson databind version.
Fixed actions URL override.
Fix add component to teams.
Change component template user to String.
Create parent directories before writing export files.
Changes to command line property handling.
Fixes to add application to teams and set user team preferences.

* Added delete authentication realm, delete authorization realm actions, get statuses actions.
Added post-processing script actions.
Added get system properties and delete system property action.
Fixed HTTP select import problems.
Updated jackson databind version.
Fixed actions URL override.
Fix add component to teams.
Change component template user to String.
Create parent directories before writing export files.
Changes to command line property handling.
Fixes to add application to teams and set user team preferences.

* Added ability to validate single full versions and to get request versions as a map/list. Changes to get resource inventory action and handling of Eval. Added fail if not found to delete resource action.

* Fix updating system configuration.
Added delete automation plugin action.
Added commit option to remove tag from agents. Added get automation plugin actions.
Fix sync ldap action and resource action tests.
Fixes for LDAP actions.

* Skip zip if extracted.
Fix download version files extract and add options.
Fixed Matcher replacement in actions runner.

* Added get generic process request properties action.
Refactored package files.
Created test actions package.
Fixed group synchronization.
Fixed run application process throw exception.
Fixed create snapshot action.
Fixed typo.
Changes/Fixes to UCD session initialization.
Made Java package loading static.
Removed unused plugin runner code.

+++ src/main/groovy/org/urbancode/ucadf/core/action/ucd/agent/UcdRemoveTagsFromAgent.groovy
@@ -18,6 +18,9 @@ class UcdRemoveTagsFromAgent extends UcAdfAction {
 	/** The list of tag names or IDs. */
 	List<String> tags
 
+	/** The flag that indicates to perform the removal, otherwise show that the removal would be done. Default is true. */
+	Boolean commit = true
+
 	/**
 	 * Runs the action.	
 	 */
@@ -27,17 +30,21 @@ class UcdRemoveTagsFromAgent extends UcAdfAction {
 		validatePropsExist()
 
 		for (tag in tags) {
-			logInfo("Remove tag [$tag] from agent [$agent].")
-			
-			WebTarget target = ucdSession.getUcdWebTarget().path("/cli/agentCLI/tag")
-				.queryParam("agent", agent)
-				.queryParam("tag", tag)
+			if (commit) {
+				logInfo("Remove tag [$tag] from agent [$agent].")
 				
-			Response response = target.request(MediaType.APPLICATION_JSON).delete()
-			if (response.getStatus() == 204) {
-				logInfo("Tag [$tag] removed from agent [$agent].")
+				WebTarget target = ucdSession.getUcdWebTarget().path("/cli/agentCLI/tag")
+					.queryParam("agent", agent)
+					.queryParam("tag", tag)
+					
+				Response response = target.request(MediaType.APPLICATION_JSON).delete()
+				if (response.getStatus() == 204) {
+					logInfo("Tag [$tag] removed from agent [$agent].")
+				} else {
+		            throw new UcdInvalidValueException(response)
+				}
 			} else {
-	            throw new UcdInvalidValueException(response)
+				logInfo("Would remove tag [$tag] from agent [$agent].")
 			}
 		}
 	}
